feat(ff4-dashboards): dashboard browser module added for hosting browsers sub project independently
