+++ projects/ff4j-dashboards/src/app/dashboard.component.html
@@ -1,1 +1,1 @@
-<p>dashboard works!</p>
+<h4>Dashboard works !!</h4>
