+++ projects/ff4j-dashboards/src/app/dashboard.module.ts
@@ -1,24 +1,11 @@
 import { Injector, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { DashboardComponent } from './dashboard.component';
-import { BrowserModule } from '@angular/platform-browser';
-import { createCustomElement } from '@angular/elements';
-import { environment } from '../environments/environment';
+import { DashboardRoutingModule } from './dashboard-routing.module';
 
 @NgModule({
   declarations: [DashboardComponent],
-  imports: [CommonModule, BrowserModule],
-  entryComponents: [DashboardComponent],
-  bootstrap: environment.loadBootstrap ? [DashboardComponent] : [],
+  imports: [CommonModule, DashboardRoutingModule],
   exports: [DashboardComponent],
 })
-export class DashboardModule {
-  constructor(private injector: Injector) {}
-
-  ngDoBootstrap() {
-    const dashboardElement = createCustomElement(DashboardComponent, {
-      injector: this.injector,
-    });
-    customElements.define('ff4j-dashboard-element', dashboardElement);
-  }
-}
+export class DashboardModule {}
